searchable measurable trees
 - done editor page
 - pulled hierarchy functions into individual file

#1377
