+++ waltz-ng/client/measurable-rating/components/tree/measurable-rating-tree.js
@@ -17,7 +17,9 @@
  */
 
 import _ from 'lodash';
-import {initialiseData, buildHierarchies, switchToParentIds} from '../../../common';
+import {initialiseData} from '../../../common';
+import {buildHierarchies, switchToParentIds, prepareSearchNodes, doSearch} from '../../../common/hierarchy-utils';
+
 
 
 /**
@@ -28,18 +30,20 @@ import {initialiseData, buildHierarchies, switchToParentIds} from '../../../comm
  *
  * Intended only for use with a single application and a single measurable kind.
  */
-
-
 const bindings = {
     ratings: '<',
     measurables: '<',
-    onSelect: '<'
+    onSelect: '<',
+    scrollHeight: '@' // should correspond to numeric values in `waltz-scroll-region` classes
 };
 
 
 const initialState = {
+    containerClass: '',
+    hierarchy: [],
     measurables: [],
     ratings: [],
+    searchTerms: '',
     treeOptions: {
         nodeChildren: "children",
         dirSelectable: true,
@@ -51,28 +55,17 @@ const initialState = {
             }
         }
     },
-    onSelect: d => console.log('default on-select for measurable-rating-tree: ', d)
+    onSelect: (m, r) => console.log('default on-select for measurable-rating-tree: ', m, r)
 };
 
 
 const template = require('./measurable-rating-tree.html');
 
 
-function enrichNodes(ratings = [], measurables = []) {
-    const ratingsByMeasurableId = _.keyBy(ratings, 'measurableId');
-    return _.map(measurables, m => ({
-        id: m.id,
-        parentId: m.parentId,
-        measurable: m,
-        rating: ratingsByMeasurableId[m.id]
-    }));
-}
-
-
 // expand nodes with a rating (incl. parents)
-function calculateExpandedNodes(nodes = []) {
+function calculateExpandedNodes(nodes = [], ratingsById = {}) {
     const byId = _.keyBy(nodes, 'id');
-    const startingNodes = _.filter(nodes, n => n.rating);
+    const startingNodes = _.filter(nodes, n => ratingsById[n.id] != null);
 
     const requiredIds = [];
     _.each(startingNodes, n => {
@@ -90,17 +83,37 @@ function calculateExpandedNodes(nodes = []) {
 }
 
 
+function prepareTree(nodes = []) {
+    return switchToParentIds(buildHierarchies(nodes));
+
+}
+
 function controller() {
     const vm = this;
 
     vm.$onInit = () => initialiseData(vm, initialState);
 
-    vm.$onChanges = () => {
-        const nodes = enrichNodes(vm.ratings, vm.measurables);
-        vm.treeData = switchToParentIds(buildHierarchies(nodes));
+    vm.$onChanges = (c) => {
+        vm.searchNodes = prepareSearchNodes(vm.measurables);
+        vm.hierarchy = prepareTree(vm.measurables);
+        vm.ratingsByMeasurable = _.keyBy(vm.ratings || [], 'measurableId');
+
         if (_.isEmpty(vm.expandedNodes)) {
-            vm.expandedNodes = calculateExpandedNodes(nodes);
+            vm.expandedNodes = calculateExpandedNodes(vm.measurables, vm.ratingsByMeasurable);
         }
+
+        if (c.scrollHeight) {
+            vm.containerClass = `waltz-scroll-region-${vm.scrollHeight}`;
+        }
+    };
+
+    vm.searchTermsChanged = (termStr = '') => {
+        vm.hierarchy = prepareTree(doSearch(termStr, vm.searchNodes));
+    };
+
+    vm.clearSearch = () => {
+        vm.searchTermsChanged('');
+        vm.searchTerms = '';
     };
 }
