+++ waltz-ng/client/measurable-rating/measurable-rating-edit.html
@@ -166,10 +166,10 @@
                     <div ng-repeat="tab in ctrl.tabs track by tab.kind.code"
                          ng-class="{ 'wt-active': tab.kind.code == ctrl.visibility.tab }"
                          class="wt-tab">
-                        <div class="waltz-scroll-region-600"
-                             style="padding-top: 6px">
+                        <div style="padding-top: 6px">
                             <waltz-measurable-rating-tree ratings="tab.ratings"
                                                           measurables="tab.measurables"
+                                                          scroll-height="600"
                                                           on-select="ctrl.onMeasurableSelect">
                             </waltz-measurable-rating-tree>
                         </div>
